# rest_client/openapi.yaml
openapi: 3.0.3
info:
  title: house-keeper-pro
  description: house-keeper-proのAPI設計
  version: 0.1.1
servers:
- url: http://127.0.0.1:3000
  description: local sam
- url: http://127.0.0.1:80
  description: local fastapi
paths:
  /lineapi/callback:
    post:
      summary: line messaging
      description: ユーザーからメッセージが送信された際、LINE Message APIからこちらのメソッドが呼び出されます。
      operationId: lineMessaging
      parameters:
      - name: x-line-signature
        in: header
        required: false
        schema:
          title: X-Line-Signature
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /report/{form_id}:
    get:
      summary: get report form
      operationId: getReportForm
      parameters:
      - name: form_id
        in: path
        required: true
        schema:
          type: integer
          title: フォームのID
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      summary: create report form
      operationId: createReportForm
      parameters:
      - name: form_id
        in: path
        required: true
        schema:
          type: integer
          title: フォームのID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnswerBase'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /report/{form_id}/answer/{answer_id}/order/{stub_id}:
    post:
      summary: create report img
      operationId: createReportImg
      parameters:
      - name: form_id
        in: path
        required: true
        schema:
          type: integer
          title: フォームのID
      - name: answer_id
        in: path
        required: true
        schema:
          type: integer
          title: 回答フォームのID
      - name: stub_id
        in: path
        required: true
        schema:
          type: integer
          title: 回答のID
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/createReportImgRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
components:
  schemas:
    AnswerBase:
      properties:
        timeInputStart:
          type: string
          format: date-time
          title: 業務開始時間
        timeInputEnd:
          type: string
          format: date-time
          title: 業務終了時間
        lineToken:
          type: string
          title: lineのユーザートークン
      type: object
      required:
      - timeInputStart
      - timeInputEnd
      - lineToken
      title: AnswerBase
    createReportImgRequest:
      properties:
        lineToken:
          type: string
          title: lineのユーザートークン
        beforeImage:
          type: string
          format: binary
          title: Beforeimage
        afterImage:
          type: string
          format: binary
          title: Afterimage
      type: object
      required:
      - lineToken
      - beforeImage
      - afterImage
      title: createReportImgRequest
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
tags:
- name: user
  description: ユーザーに関するAPI
- name: seal
  description: スタンプを押す用のAPIとか